Project Foundation. - Models restructure
