+++ app/src/main/java/com/example/themealsapp/model/meal/MealResponse.kt
@@ -1,11 +1,11 @@
-package com.example.themealsapp.model
+package com.example.themealsapp.model.meal
 
 
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
-data class IngredientResponse(
+data class MealResponse(
     @Json(name = "meals")
-    val meals: List<Ingredient>? = listOf()
+    val meals: List<Meal>? = listOf()
 )
