+++ app/src/main/java/com/example/themealsapp/rest/MealsAPI.kt
@@ -1,6 +1,6 @@
 package com.example.themealsapp.rest
 
-import com.example.themealsapp.model.MealResponse
+import com.example.themealsapp.model.meal.MealResponse
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Query
@@ -11,7 +11,7 @@ interface MealsAPI {
     suspend fun searchMealById(
         @Query("i") i: String
     ): Response<MealResponse>
-
+    // www.themealdb.com/api/json/v1/1/lookup.php?i=52772
     companion object {
         const val BASE_URL = "https://www.themealdb.com/api/json/v1/1/"
         private const val FILTER = "filter.php"
