+++ app/src/main/java/com/example/themealsapp/rest/MealsRepository.kt
@@ -1,6 +1,6 @@
 package com.example.themealsapp.rest
 
-import com.example.themealsapp.model.MealResponse
+import com.example.themealsapp.model.meal.MealResponse
 import com.example.themealsapp.utils.UIState
 import kotlinx.coroutines.flow.Flow
 
@@ -9,11 +9,11 @@ interface MealsRepository {
     /**
      * Get a meal by providing its ID
      */
-    fun getMealById(): Flow<UIState<MealResponse>>
+    fun getMealById(i : String): Flow<UIState<MealResponse>>
 }
 
 class MealsRepositoryImpl() : MealsRepository {
-    override fun getMealById(): Flow<UIState<MealResponse>> {
+    override fun getMealById(i : String): Flow<UIState<MealResponse>> {
         TODO("Not yet implemented")
     }
 }
