+++ app/src/main/java/com/example/themealsapp/utils/CustomExceptions.kt
@@ -1,4 +1,4 @@
 package com.example.themealsapp.utils
 
-class NullMealsResponse(message: String = "Schools response is null") : Exception(message)
+class NullMealsResponse(message: String = "Meals response is null") : Exception(message)
 class FailureResponse(message: String?) : Exception(message)
